refactor(scene3d): 抽出检查器面板到 scene3dInspector.tsx [split 2/N]
VectorInputs/ColorField/SceneObjectList/MannequinPosePanel/PropertyPanel 移出;
共用纯 helper mannequinRoleLabel 归 scene3dMath(避循环依赖);壳 3478→2815。
双 tsc 零错误,scene3d 测试绿
